+++ gryannote/audio/frontend/player/AudioPlayerWithAnnotation.svelte
@@ -305,6 +305,28 @@
 	};
 
 	/**
+	 * Return whether specified region is visible on the screen
+	 * @param region
+	 * @param waveformContainer
+	 */
+	function isRegionVisible(region: Region, waveformContainer: HTMLDivElement): bool {
+		// Get the left and right boundaries of the waveform view box:
+		const viewbox = waveformContainer.getBoundingClientRect();
+		const viewboxLeft = viewbox.left;
+		const viewboxRight = viewbox.right;
+
+		// Get the left and right boundaries of the region
+		const regionBox = region.element.getBoundingClientRect();
+		const regionBoxLeft = regionBox.left;
+		const regionBoxRight = regionBox.right;
+
+		// Check if the region is within the visible bounds of the visible part of the waveform
+		const isVisible = (regionBoxLeft >= viewboxLeft && regionBoxRight <= viewboxRight);
+
+    	return isVisible;
+	}
+
+	/**
 	 * Set active region with specified one.
 	 * Do nothing if component is not in interactive mode
 	 * @param region the region to activate
@@ -324,7 +346,13 @@
 		setActiveRegionBackground(region.color);
 
 		// adjust time cursor position so that active region is always visible
-		waveform.setTime(region.start)
+		// only update if audio is not playing and active region not visible
+		// to prevent audio jumping
+		if(!waveform.isPlaying()){
+			if(!isRegionVisible(region, container)){
+				waveform.setTime(region.start);
+			}
+		}
 	}
 
 	/**