+++ CHANGELOG.md
@@ -16,6 +16,8 @@ demo = gr.Interface(lambda x : x, inputs=None, outputs=player)
 ```
 
 ### improvements
+- if audio is playing, time cursor will not jump to the start of the active annotation
+- if audio is paused, time cursor will be set to the start of the active annotation only if this annotation is not visible on the screen.
 - replace arithmetic zoom (z = z + delta) by a geometric one (z = z * coef)
 - add new shortcuts to speed up zoom in / zoom out: `SHIFT+UP` / `SHIFT+DOWN`